check if load image

+++ MyApplicationVolume/app/src/main/java/com/tracingfunc/app2/MyMarker.java
@@ -403,10 +403,10 @@ public class MyMarker implements Cloneable{
         return outMask;
     }
 
-    public static int[] swcToMask2(Vector<MyMarker> inswc, int[] sz, int flag){
+    public static byte[] swcToMask2(Vector<MyMarker> inswc, int[] sz, int flag){
         int totalSz = sz[0]*sz[1]*sz[2];
         int sz01 = sz[0]*sz[1];
-        int[] outMask = new int[totalSz];
+        byte[] outMask = new byte[totalSz];
         for(int i=0; i<totalSz; i++)
             outMask[i] = 0;
         Vector<MyMarker> leafMarkers = HierarchyPruning.getLeafMarkers(inswc);
@@ -414,7 +414,7 @@ public class MyMarker implements Cloneable{
         for(int i=0; i<leafMarkers.size(); i++){
             MyMarker leaf = leafMarkers.get(i);
             MyMarker p = leaf;
-            outMask[(int) p.z*sz01+(int) p.y*sz[0]+(int) p.x] = flag;
+            outMask[(int) p.z*sz01+(int) p.y*sz[0]+(int) p.x] = (byte) flag;
             while (!visitedMarkers.contains(p) && p.parent != null){
                 MyMarker par = p.parent;
                 if(MyMarker.dist(par,p)>1){
@@ -423,9 +423,9 @@ public class MyMarker implements Cloneable{
                     int z = (int) ((p.z+par.z)/2);
                     if(x<0 || x>=sz[0] || y<0 || y>=sz[1] ||z<0 || z>=sz[2])
                         break;
-                    outMask[z*sz01+y*sz[0]+x] = flag;
+                    outMask[z*sz01+y*sz[0]+x] = (byte) flag;
                 }
-                outMask[(int) par.z*sz01+(int) par.y*sz[0]+(int) par.x] = flag;
+                outMask[(int) par.z*sz01+(int) par.y*sz[0]+(int) par.x] = (byte) flag;
                 visitedMarkers.add(p);
                 p = par;
             }
